feat(server): allow recommended fps to be a float

+++ src/common/models/stream_video.py
@@ -29,4 +29,4 @@ class VideoStreamRequestMessage:
 class VideoStreamResponseMessage:
   """Implements the VideoStream response message arguments."""
 
-  recommended_fps: int
+  recommended_fps: float
